* Fix issue where blender animation importer was truncating the animation duration to an integer (thanks rectalogic)
git-svn-id: https://jmonkeyengine.googlecode.com/svn/trunk@9626 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ engine/src/blender/com/jme3/scene/plugins/blender/modifiers/ObjectAnimationModifier.java
@@ -50,7 +50,7 @@ import java.util.logging.Logger;
 		// calculating track
 		SpatialTrack track = (SpatialTrack) ipo.calculateTrack(-1, 0, ipo.getLastFrame(), fps, true);
 
-		Animation animation = new Animation(objectAnimationName, ipo.getLastFrame() / fps);
+		Animation animation = new Animation(objectAnimationName, ipo.getLastFrame() / (float)fps);
 		animation.setTracks(new SpatialTrack[] { track });
 		ArrayList<Animation> animations = new ArrayList<Animation>(1);
 		animations.add(animation);
